allow Wh and Ah as names
I think I have a core way to fix this that is less frustrating
and far more structured than raw Python code, by using
a JSON file or some other custom but less terse format that allows for
a reader to make use of `eval` and `setattr`,
and pickle the result of all analysis to cache
so that it is less frustrating for startup.

This seems like a more mature model for in the future.